Added update log page
