+++ package.json
@@ -15,6 +15,7 @@
     "vite": "^5.4.1"
   },
   "dependencies": {
-    "flexsearch": "^0.7.43"
+    "flexsearch": "^0.7.43",
+    "svelte-spa-router": "^4.0.1"
   }
 }
